feat: add routing with react-routing-dom
