+++ frontend/src/App.js
@@ -1,26 +1,39 @@
-import React, { useState, useContext } from "react";
+import React, { useState } from "react";
+import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
+
 import { 
-  Button, 
-  Card,
-  CardHeader,
-  CardBody,
-  CardFooter,
-  Grid,
+  Button,
   Grommet, 
-  grommet,
+  grommet as grommetTheme,
   Header, 
-  Heading,
-  Page, 
-  PageContent, 
-  PageHeader, 
-  Paragraph,
-  ResponsiveContext,
+  Page,
   Text 
 } from 'grommet';
 import { deepMerge } from "grommet/utils";
 import { Moon, Sun } from "grommet-icons";
 
-const theme = deepMerge(grommet, {
+import { Home } from './pages';
+import { AppBar } from './components';
+
+const pages = [
+  // {
+  //   label: "Servers",
+  //   Icon: ServersIcon,
+  //   path: "/servers"
+  // },
+  // {
+  //   label: "Users",
+  //   Icon: Group,
+  //   path: "/users"
+  // },
+  // {
+  //   label: "Settings",
+  //   Icon: SettingsOption,
+  //   path: "/settings"
+  // }
+];
+
+const theme = deepMerge(grommetTheme, {
   global: {
     colors: {
       brand: '#228BE6',
@@ -33,65 +46,28 @@ const theme = deepMerge(grommet, {
   },
 });
 
-const AppBar = (props) => (
-  <Header
-    background="brand"
-    pad={{ left: "medium", right: "small", vertical: "small" }}
-    elevation="medium"
-    {...props}
-  />
-);
-
-const CardTemplate = ({ title }) => {
-  const size = useContext(ResponsiveContext);
-  return (
-    <Card>
-      <CardHeader pad="medium">
-        <Heading level={2} margin="none">
-          {title}
-        </Heading>
-      </CardHeader>
-      <CardBody pad="medium">
-        <Paragraph maxLines={size === "small" ? 3 : undefined}>
-          Lorem ipsum dolor sit amet, consectetur adipiscing elit. Maecenas
-          porttitor non nulla ac vehicula. Aliquam erat volutpat. Mauris auctor
-          faucibus est at mattis. Aliquam a enim ac nisi aliquam consectetur et
-          ac velit. Mauris ut imperdiet libero.
-        </Paragraph>
-      </CardBody>
-      <CardFooter pad="medium" background="background-contrast">
-        Footer
-      </CardFooter>
-    </Card>
-  );
-};
-  
-
 function App() {
   const [dark, setDark] = useState(false);
 
   return (
-    <Grommet theme={theme} full themeMode={dark ? "dark" : "light"}>
-      <Page>
+    <Router>
+      <Grommet theme={theme} full themeMode={dark ? "dark" : "light"}>
+        <Page>
 
-        <AppBar>
-          <Text size="large">Activitn</Text>
-          <Button
-            a11yTitle={dark ? "Switch to Light Mode" : "Switch to Dark Mode"}
-            icon={dark ? <Moon /> : <Sun />}
-            onClick={() => setDark(!dark)}
-          />
-        </AppBar>
-        <PageContent>
-          <PageHeader title="Welcome to Grommet!" />
-          <Grid columns="medium" gap="large" pad={{ bottom: "large" }}>
-            <CardTemplate title={"Card 1"} />
-            <CardTemplate title={"Card 2"} />
-            <CardTemplate title={"Card 3"} />
-          </Grid>
-        </PageContent>
-      </Page>
-    </Grommet>
+          <AppBar>
+            <Text size="large">Activitn</Text>
+            <Button
+              a11yTitle={dark ? "Switch to Light Mode" : "Switch to Dark Mode"}
+              icon={dark ? <Moon /> : <Sun />}
+              onClick={() => setDark(!dark)}
+            />
+          </AppBar>
+          <Routes>
+            <Route index element={<Home />} />
+          </Routes>
+        </Page>
+      </Grommet>
+    </Router>
   );
 }
