+++ frontend/src/components/AppBar.js
@@ -1,0 +1,12 @@
+import {
+    Header, 
+} from 'grommet';
+
+export const AppBar = (props) => (
+<Header
+    background="brand"
+    pad={{ left: "medium", right: "small", vertical: "small" }}
+    elevation="medium"
+    {...props}
+/>
+);
