+++ frontend/src/pages/index.js
@@ -1,0 +1,2 @@
+
+export * from './Home';
